Separating control over auto deletion to a new setting

+++ includes/class-apple-news.php
@@ -79,6 +79,32 @@ class Apple_News {
 	}
 
 	/**
+	 * Validate settings and see if any updates need to be performed.
+	 *
+	 * @param array $wp_settings An array of settings loaded from WP options.
+	 *
+	 * @access public
+	 * @return array The modified settings array.
+	 */
+	public function validate_settings( $wp_settings ) {
+		// If this option doesn't exist, either the site has never installed
+		// this plugin or they may be using an old version with individual
+		// options. To be safe, attempt to migrate values. This will happen only
+		// once.
+		if ( false === $wp_settings ) {
+			$wp_settings = $this->migrate_settings( $wp_settings );
+		}
+
+		// Check for presence of legacy header settings and migrate to new.
+		$wp_settings = $this->migrate_header_settings( $wp_settings );
+
+		// Check for presence of legacy API settings and migrate to new.
+		$wp_settings = $this->migrate_api_settings( $wp_settings );
+
+		return $wp_settings;
+	}
+
+	/**
 	 * Migrate legacy header settings to new format.
 	 *
 	 * @param array $wp_settings An array of settings loaded from WP options.
@@ -136,6 +162,27 @@ class Apple_News {
 	}
 
 	/**
+	 * Initialize the value of api_autosync_delete if not set.
+	 *
+	 * @param array $wp_settings An array of settings loaded from WP options.
+	 *
+	 * @access public
+	 * @return array The modified settings array.
+	 */
+	public function migrate_api_settings( $wp_settings ) {
+		// Use the value of api_autosync_update for api_autosync_delete if not set
+		// since that was the previous value used to determine this behavior.
+		if ( empty( $wp_settings['api_autosync_delete'] )
+			&& ! empty( $wp_settings['api_autosync_update'] ) ) {
+
+			$wp_settings['api_autosync_delete'] = $wp_settings['api_autosync_update'];
+			update_option( self::$option_name, $wp_settings, 'no' );
+		}
+
+		return $wp_settings;
+	}
+
+	/**
 	 * Attempt to migrate settings from an older version of this plugin
 	 *
 	 * @param Settings $settings
@@ -146,16 +193,21 @@ class Apple_News {
 		// For each potential value, see if the WordPress option exists.
 		// If so, migrate its value into the new array format.
 		// If it doesn't exist, just use the default value.
-		foreach ( $settings->all() as $key => $default ) {
-			$value = get_option( $key, $default );
-			$migrated_settings[ $key ] = $value;
+		if ( is_object( $settings ) ) {
+			$all_settings = $settings->all();
+			if ( ! empty( $all_settings ) && is_array( $all_settings ) ) {
+				foreach ( $all_settings as $key => $default ) {
+					$value = get_option( $key, $default );
+					$migrated_settings[ $key ] = $value;
+				}
+
+				// Store these settings
+				update_option( self::$option_name, $migrated_settings, 'no' );
+
+				// Delete the options to clean up
+				array_map( 'delete_option', array_keys( $migrated_settings ) );
+			}
 		}
-
-		// Store these settings
-		update_option( self::$option_name, $migrated_settings, 'no' );
-
-		// Delete the options to clean up
-		array_map( 'delete_option', array_keys( $migrated_settings ) );
 
 		return $migrated_settings;
 	}
